ai_career_assistant/requirements.txt
django==4.2.7
djangorestframework==3.14.0
openai==1.10.0
pinecone-client==2.2.2
supabase==1.0.3
mixpanel==4.10.0
python-dotenv==1.0.0
requests==2.31.0
